requirements.txt
click==8.1.3
numpy==1.24.1
pandas==1.5.3
psycopg2==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.0
typing_extensions==4.4.0
